we display our product list offline

+++ app/src/main/java/com/rakutentest/android/data/repository/BuyBoxRepositoryImpl.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.Flow
 class BuyBoxRepositoryImpl(
     private val buyBoxLocalDataSource: BuyBoxLocalDataSource
 ): BuyBoxRepository {
-    override fun getBuyBox(productId: Int): Flow<BuyboxRoom> {
+    override fun getBuyBox(productId: Long): Flow<BuyboxRoom> {
         return buyBoxLocalDataSource.getLocalBuyBox(productId = productId)
     }
