+++ app/src/main/java/com/rakutentest/android/data/repository/dataSource/buyBox/BuyBoxLocalDataSource.kt
@@ -5,7 +5,7 @@ import com.rakutentest.android.data.model.dataLocal.ProductRoom
 import kotlinx.coroutines.flow.Flow
 
 interface BuyBoxLocalDataSource {
-    fun getLocalBuyBox(productId: Int): Flow<BuyboxRoom>
+    fun getLocalBuyBox(productId: Long): Flow<BuyboxRoom>
 
     suspend fun insertBuyBox(buyBox: BuyboxRoom)
 }
