+++ app/src/main/java/com/rakutentest/android/domain/repository/BuyBoxRepository.kt
@@ -5,7 +5,7 @@ import com.rakutentest.android.data.model.dataLocal.ProductRoom
 import kotlinx.coroutines.flow.Flow
 
 interface BuyBoxRepository {
-    fun getBuyBox(productId: Int): Flow<BuyboxRoom>
+    fun getBuyBox(productId: Long): Flow<BuyboxRoom>
 
     suspend fun insertBuyBox(buyBox: BuyboxRoom)
 }
